api/requirements.txt
fastapi>=0.85.0
uvicorn>=0.19.0
torch>=2.0.0
torchvision>=0.15.0
pillow>=9.4.0
python-multipart>=0.0.6
einops
pytest
httpx
google-cloud-storage